agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Voice agent for I-Mart using Pipecat and LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pipecat-ai[deepgram,openai,silero]>=0.0.54",
    "aiortc>=1.9.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "websockets>=13.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.6.0",
    "aiohttp>=3.11.0",
    "redis>=5.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
